chore: fix tiangle color and delete template
